[FIX] Ao fazer o pull, confere se o arquivo sendo extraído do zip já existe na pasta locked. Se não existir, salva. Antes, o comando Files.copy estava causando erro ao tentar salvar um arquivo que já existia na pasta e estava interrompendo a execução do app.

+++ src/main/java/fateczl/CriptoGitClient/service/PullService.java
@@ -104,26 +104,28 @@ public class PullService {
             
             while ((entry = zipInputStream.getNextEntry()) != null) {
                 Path entryPath = lockedDir.resolve(entry.getName());
-                
-                // Previne zip slip attack
-                if (!entryPath.normalize().startsWith(lockedDir.normalize())) {
-                    throw new Exception("Entrada inválida no zip: " + entry.getName());
-                }
-                
-                if (entry.isDirectory()) {
-                    // Cria o diretório
-                    Files.createDirectories(entryPath);
-                } else {
-                    // Cria os diretórios pais se necessário
-                    if (entryPath.getParent() != null) {
-                        Files.createDirectories(entryPath.getParent());
+
+                if (!Files.exists(entryPath)) {                
+                    // Previne zip slip attack
+                    if (!entryPath.normalize().startsWith(lockedDir.normalize())) {
+                        throw new Exception("Entrada inválida no zip: " + entry.getName());
                     }
                     
-                    // Escreve o arquivo
-                    Files.copy(zipInputStream, entryPath);
+                    if (entry.isDirectory()) {
+                        // Cria o diretório
+                        Files.createDirectories(entryPath);
+                    } else {
+                        // Cria os diretórios pais se necessário
+                        if (entryPath.getParent() != null) {
+                            Files.createDirectories(entryPath.getParent());
+                        }
+                        
+                        // Escreve o arquivo
+                        Files.copy(zipInputStream, entryPath);
+                    }
+                    
+                    zipInputStream.closeEntry();
                 }
-                
-                zipInputStream.closeEntry();
             }
         }
     }
